Add TestRule to DiskSchedulerTest for debug
Under the original code, debugging the test file with SHIFT-CMD-F6 would cause the tests to fail, as the catch-all timeout rule would be triggered. This modifies the rule such that normal testing of the file will trigger a timeout exception after 1000 milliseconds, but running the code in debug (with breakpoints, etc) will not.
This uses junit's DisableOnDebug with a global TestRule defined, which replaces the @Test(timeout=1000).